refactor: Extract the shared behaviors so they can be used by collections tests.
